Move AddToProject to AppStack
