+++ src/frontend/screens/Onboarding/index.ts
@@ -1,4 +1,3 @@
-export * from "./AddToProjectScreen";
 export * from "./CreateOrJoinScreen";
 export * from "./JoinProjectQrScreen";
 export * from "./SendJoinRequestScreen";
